medtrack-analytics/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
pandas==2.0.3
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.0
pytest-flask==1.2.0
faker==19.3.1
Werkzeug==2.3.7
